Title: 최대 숫자 만들기, Time: 702ms, Memory: 176MB, Status: Memory Limit Exceed - CodeTree

+++ 240827/최대 숫자 만들기/make-biggest-num.java
@@ -7,27 +7,27 @@ public class Main {
 
         int N = Integer.parseInt(br.readLine());
 
-        Integer[] nums = new Integer[N];
+        Long[] nums = new Long[N];
 
-        for (int i=0; i<N; i++) nums[i] = Integer.parseInt(br.readLine());
+        for (int i=0; i<N; i++) nums[i] = Long.parseLong(br.readLine());
 
-        Arrays.sort(nums, new Comparator<Integer>() {
+        Arrays.sort(nums, new Comparator<Long>() {
             @Override
-            public int compare(Integer a, Integer b) {
-                String aS = Integer.toString(a);
-                String bS = Integer.toString(b);
+            public int compare(Long a, Long b) {
+                String aS = Long.toString(a);
+                String bS = Long.toString(b);
                 
-                int ab = Integer.parseInt(aS + bS);
-                int ba = Integer.parseInt(bS + aS);
+                long ab = Long.parseLong(aS + bS);
+                long ba = Long.parseLong(bS + aS);
 
                 // 내림차순으로 정렬
-                return Integer.compare(ba, ab);
+                return Long.compare(ba, ab);
             }
         });
 
         String ans = "";
-        for (int num : nums) {
-            ans += Integer.toString(num);
+        for (long num : nums) {
+            ans += Long.toString(num);
         }
 
         System.out.println(ans);
